Update epaper interface.
Ditch the abstract base class because we are using templates
and do not inherit from this. If concrete display class does not have
needed methods and variables it would not compile.

The interface is designed to take color in RGB565 format for greater
compatibility. Base colors in RGB565 are defined.

Cleanup.