Add a gfx wrapper.
Any display from laskakit_epaper can be now wrapped
in our gfx class that provides access to Adafruit_GFX.
Displays can be then controlled that way which is very easy.

Also if one does not want use it at all, then simply do not include the
header and the code won't be even compiled to the resulting library.